+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 CHANGELOG
 =========
+1.0.2
+-----
+* Fixed the bug #23: "Unable to parse ArrayTables that contain Tables".
+* A new class `KeyStore` has been added to deal with the logic of the keys (keys, tables and array of tables).
+
 1.0.1 (2018-02-05)
 ------------------
 * Fixed a bug related to integer keys: now, it's possible to create keys using an integer. Reported by @betrixed.